refactor : GameUser 위치변경
