+++ src/main/java/bowling/domain/GameUser.java
@@ -1,4 +1,4 @@
-package bowling.vo;
+package bowling.domain;
 
 import java.util.regex.Pattern;
