+++ src/test/java/bowling/vo/GameUserTest.java
@@ -1,5 +1,6 @@
 package bowling.vo;
 
+import bowling.domain.GameUser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
